Create plugin loaders to test TensorFlow install
