+++ tensorboard/BUILD
@@ -155,18 +155,18 @@ py_library(
     deps = [
         "//tensorboard/plugins:base_plugin",
         "//tensorboard/plugins/audio:audio_plugin",
-        "//tensorboard/plugins/beholder:beholder_plugin",
+        "//tensorboard/plugins/beholder:beholder_plugin_loader",
         "//tensorboard/plugins/core:core_plugin",
         "//tensorboard/plugins/custom_scalar:custom_scalars_plugin",
         "//tensorboard/plugins/debugger:debugger_plugin_loader",
         "//tensorboard/plugins/distribution:distributions_plugin",
         "//tensorboard/plugins/graph:graphs_plugin",
         "//tensorboard/plugins/histogram:histograms_plugin",
-        "//tensorboard/plugins/hparams:hparams_plugin",
+        "//tensorboard/plugins/hparams:hparams_plugin_loader",
         "//tensorboard/plugins/image:images_plugin",
-        "//tensorboard/plugins/interactive_inference:interactive_inference_plugin",
+        "//tensorboard/plugins/interactive_inference:interactive_inference_plugin_loader",
         "//tensorboard/plugins/pr_curve:pr_curves_plugin",
-        "//tensorboard/plugins/profile:profile_plugin",
+        "//tensorboard/plugins/profile:profile_plugin_loader",
         "//tensorboard/plugins/projector:projector_plugin",
         "//tensorboard/plugins/scalar:scalars_plugin",
         "//tensorboard/plugins/text:text_plugin",
