+++ tensorboard/plugins/hparams/BUILD
@@ -40,6 +40,18 @@ py_library(
     ],
 )
 
+py_library(
+    name = "hparams_plugin_loader",
+    srcs = ["hparams_plugin_loader.py"],
+    srcs_version = "PY2AND3",
+    deps = [
+        ":hparams_plugin",
+        "//tensorboard/plugins:base_plugin",
+        "//tensorboard/util:tb_logging",
+        "@org_pythonhosted_six",
+    ],
+)
+
 py_test(
     name = "list_session_groups_test",
     size = "small",
