+++ tensorboard/plugins/interactive_inference/BUILD
@@ -17,6 +17,18 @@ py_library(
     ],
 )
 
+py_library(
+    name = "interactive_inference_plugin_loader",
+    srcs = ["interactive_inference_plugin_loader.py"],
+    srcs_version = "PY2AND3",
+    deps = [
+        ":interactive_inference_plugin",
+        "//tensorboard/plugins:base_plugin",
+        "//tensorboard/util:tb_logging",
+        "@org_pythonhosted_six",
+    ],
+)
+
 py_test(
     name = "interactive_inference_plugin_test",
     size = "small",
